requirements.txt
# Core Framework
fastapi
uvicorn[standard]
websockets
python-multipart
pydantic
pydantic-settings

# Database
sqlalchemy
alembic
psycopg2-binary

# LLM - Cerebras SDK
cerebras-cloud-sdk

# STT - Deepgram
deepgram-sdk

# TTS - ElevenLabs
elevenlabs
pydub
audioop-lts  # Backport of audioop for Python 3.13+

# Vector DB - FAISS
faiss-cpu

# Embeddings - HuggingFace SentenceTransformers
sentence-transformers

# Twilio
twilio

# Utilities
python-dotenv
httpx
aiofiles

# Development & Testing (optional but recommended)
pytest
pytest-asyncio
black
ruff
